move startPublicThreadWithMessage to thread parents
