+++ core/src/main/kotlin/behavior/channel/threads/ThreadParentChannelBehavior.kt
@@ -2,6 +2,7 @@ package dev.kord.core.behavior.channel.threads
 
 import dev.kord.common.entity.ArchiveDuration
 import dev.kord.common.entity.ChannelType
+import dev.kord.common.entity.Snowflake
 import dev.kord.common.entity.optional.Optional
 import dev.kord.core.behavior.channel.GuildMessageChannelBehavior
 import dev.kord.core.cache.data.ChannelData
@@ -28,6 +29,20 @@ interface ThreadParentChannelBehavior : GuildMessageChannelBehavior {
 
     suspend fun startPublicThread(name: String, archiveDuration: ArchiveDuration): ThreadChannel
 
+
+    suspend fun startPublicThreadWithMessage(
+        messageId: Snowflake,
+        name: String,
+        archiveDuration: ArchiveDuration
+    ): ThreadChannel {
+
+        val response = kord.rest.channel.startPublicThread(id, messageId, StartThreadRequest(name, archiveDuration))
+        val data = ChannelData.from(response)
+
+        return Channel.from(data, kord) as ThreadChannel
+    }
+
+
 }
 
 /**
@@ -52,7 +67,6 @@ interface PrivateThreadParentChannelBehavior : ThreadParentChannelBehavior {
     }
 
     suspend fun startPrivateThread(name: String, archiveDuration: ArchiveDuration): ThreadChannel
-
 }
